+++ lib/common/test/unit-tests/services/help-service.ts
@@ -20,23 +20,6 @@ const createTestInjector = (opts?: { isProjectTypeResult: boolean; isPlatformRes
 	injector.register("logger", CommonLoggerStub);
 
 	opts = opts || { isPlatformResult: true, isProjectTypeResult: true };
-
-	injector.register("dynamicHelpProvider", {
-		getLocalVariables: () => ({})
-	});
-
-	injector.register("dynamicHelpService", {
-		isProjectType: (...args: string[]): boolean => opts.isProjectTypeResult,
-		isPlatform: (...args: string[]): boolean => { return opts.isPlatformResult; },
-		getLocalVariables: (): IDictionary<any> => {
-			const localVariables: IDictionary<any> = {};
-			localVariables["myLocalVar"] = opts.isProjectTypeResult;
-			localVariables["isLinux"] = opts.isPlatformResult;
-			localVariables["isWindows"] = opts.isPlatformResult;
-			localVariables["isMacOS"] = opts.isPlatformResult;
-			return localVariables;
-		}
-	});
 	injector.register("microTemplateService", MicroTemplateService);
 	injector.register("helpService", HelpService);
 	injector.register("opener", {
@@ -53,6 +36,15 @@ const createTestInjector = (opts?: { isProjectTypeResult: boolean; isPlatformRes
 
 	injector.registerCommand("foo", {});
 
+	const microTemplateService: any = injector.resolve("microTemplateService");
+	microTemplateService.getLocalVariables = (): IDictionary<any> => {
+		const localVariables: IDictionary<any> = {};
+		localVariables["myLocalVar"] = opts.isProjectTypeResult;
+		localVariables["isLinux"] = opts.isPlatformResult;
+		localVariables["isWindows"] = opts.isPlatformResult;
+		localVariables["isMacOS"] = opts.isPlatformResult;
+		return localVariables;
+	};
 	return injector;
 };
