--- lib/dynamic-help-provider.ts
@@ -1,15 +1,0 @@
-import * as constants from './constants';
-
-export class DynamicHelpProvider implements IDynamicHelpProvider {
-	public isProjectType(args: string[]): boolean {
-		return true;
-	}
-
-	public getLocalVariables(options: { isHtml: boolean }): IDictionary<any> {
-		const localVariables: IDictionary<any> = {
-			constants: constants
-		};
-		return localVariables;
-	}
-}
-$injector.register("dynamicHelpProvider", DynamicHelpProvider);
